test buffer.

+++ PicMergeToPdf/IMerger.cs
@@ -9,7 +9,7 @@ namespace PicMerge {
 		/// <summary>
 		/// 内存映射文件设定的最大大小。
 		/// </summary>
-		internal const long MapFileSize = 0x04000000;
+		internal const int MapFileSize = 0x04000000;
 
 		public readonly struct FileResult(uint _c, string _file, string _desc = "Success.") {
 			public readonly uint code = _c;
